Delete useless data interfaces in egretProperties
