+++ src/extension/assetsmanager/src/core/ResourceManager.ts
@@ -91,7 +91,7 @@ module RES {
 
     }
     /**
-    * @internal
+    * @private
     */
     export class ResourceManagerError extends Error {
