+++ tools/project/ProjectData.ts
@@ -124,23 +124,6 @@ class EgretProjectData {
         return [];
     }
 
-    getExmlRoots(): string[] {
-        if (globals.hasKeys(this.egretProperties, ["eui", "exmlRoot"])) {
-            let result = this.egretProperties.eui.exmlRoot;
-            if (typeof result == "string") {
-                return [_path.join(egret.args.projectDir, result)];
-            }
-            else {
-                let temp: string[] = <string[]>this.egretProperties.eui.exmlRoot;
-                return temp.reduce(function (previousValue: string[], currentValue: string) {
-                    previousValue.push(_path.join(egret.args.projectDir, currentValue));
-                    return previousValue;
-                }, []);
-            }
-        }
-        return [egret.args.projectDir];
-    }
-
     getCurrentTarget() {
         if (globals.hasKeys(this.egretProperties, ["target", "current"])) {
             return this.egretProperties.target.current;
@@ -148,20 +131,6 @@ class EgretProjectData {
         else {
             return "web"
         }
-    }
-
-    getThemes(): string[] {
-        if (globals.hasKeys(this.egretProperties, ["eui", "themes"])) {
-            return this.egretProperties.eui.themes;
-        }
-        return null;
-    }
-
-    getExmlPublishPolicy(): string {
-        if (globals.hasKeys(this.egretProperties, ["eui", "exmlPublishPolicy"])) {
-            return this.egretProperties.eui.exmlPublishPolicy;
-        }
-        return "content";
     }
 
     getCopyExmlList(): Array<string> {
